fix: Handle comments and removed use_ray_cluster config
